WebXR code clean up.
