--- src/renderers/webvr/WebVRUtils.d.ts
@@ -1,3 +1,0 @@
-import { Camera } from '../../cameras/Camera';
-
-export function setProjectionFromUnion( camera: Camera, cameraL: Camera, cameraR: Camera ): void;
