+++ src/renderers/webxr/WebXRManager.d.ts
